requirements.txt
dash==2.16.1
dash-bootstrap-components==1.5.0
pandas==2.2.1
numpy==1.26.4
plotly==5.20.0
requests==2.31.0
python-dateutil==2.9.0.post0
